requirements.txt
flask
pymongo
python-dotenv
flask_cors
jsonschema
requests_oauthlib
flask_login
authlib